modified:   main.c

+++ main.c
@@ -260,8 +260,8 @@ static void frame_clicked(GtkWidget *widget, GdkEventButton *event, gpointer use
 	gtk_widget_set_visible(GTK_WIDGET(new_image), !gtk_widget_is_visible(GTK_WIDGET(new_image)));
 	
 	gtk_fixed_move(GTK_FIXED(fixed), GTK_WIDGET(new_image), 
-			x_coord + ( canvas_winW - frame_winW ) / 2,
-			y_coord - ( canvas_winH - frame_winH ) / 2 );	
+			x_coord + (gdouble) ( canvas_winW - frame_winW ) / 2,
+			y_coord - (gdouble) ( canvas_winH - frame_winH ) / 2 );	
 	
 	g_print("click: (%lf, %lf)\n", x_coord, y_coord);
